requirements.txt
streamlit
pdfplumber
sentence-transformers
faiss-cpu
openai    # or Gemini API client if different